Rewrote code, usefull files temporary deleted
